requirements.txt
openai==1.51.2
pypdf==5.0.1
python-dotenv==1.0.1
sentence_transformers==2.7.0
torch>=2.1.2
click==8.1.8
nltk==3.9.1
absl-py==2.1.0
rouge-score==0.1.2
six==1.17.0
